+++ packages/plugin-icestark/src/lifecyclePlugin.ts
@@ -2,33 +2,18 @@ import type { Plugin } from 'vite';
 import * as babel from '@babel/core';
 import type { ParserPlugin } from '@babel/parser';
 
-export interface Entries {
-  [index: string]: string | string[];
-}
-
 /**
  * Webpack's entry would be { index: ['react-dev-utils/webpackHotDevClient.js', '/src/app''] }
  * in dev mode.
  */
-const lifecyclePlugin = (entries: Entries): Plugin => {
-  // Turn vite input to js files
-  const entryNames = Object.keys(entries)
-    .reduce((pre, next) => {
-      return [
-        ...pre,
-        // @ts-ignore
-        ...(Array.isArray(entries[next]) ? entries[next] : [entries[next]])
-      ];
-    }, [])
-    // Remove webpack hot dev client in dev
-    .filter(entry => !entry.includes('react-dev-utils/webpackHotDevClient'));
+const lifecyclePlugin = (entry: string): Plugin => {
 
   return ({
     name: 'vite-plugin-icestark-lifecycle',
     enforce: 'pre',
 
     async transform(code, id) {
-      const isEntryFile = entryNames.some((name: string) => id.includes(name));
+      const isEntryFile = id.includes(entry);
 
       if (!isEntryFile) {
         return;