+++ packages/plugin-ice-ssr/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.0.5
+
+- [fix] update dependencies version
+- [feat] support no router mode ssr
+
 ## 3.0.4
 
 - [fix] compatible with special charts in html content